Use template hook in Dashboard + minor restyle
